vfr hud working

+++ soft/firmware/src/kinematic/airspeed.c
@@ -69,7 +69,7 @@ float calc_air_speed(uint16_t press_diff_raw){
   uint16_t p;
   p = zerocomp(press_diff_raw, (int32_t)comp_data.temp_onboard);
   p = ((p * Radc) / Smpx) / KU; /* давление в паскалях */
-  return sqrtf(2*p / 1.2);
+  return sqrtf((float)(2*p) / 1.2);
 }
